build: replace to emotion
