+++ package.json
@@ -15,13 +15,16 @@
   },
   "dependencies": {
     "@docusaurus/lqip-loader": "^2.0.0-alpha.64",
+    "@emotion/css": "^11.0.0-next.17",
+    "@emotion/react": "^11.0.0-next.17",
+    "@emotion/server": "^11.0.0-next.17",
+    "@emotion/styled": "^11.0.0-next.17",
     "@mdx-js/loader": "^1.6.18",
     "@next/mdx": "^9.5.3",
     "@next/plugin-google-analytics": "^9.5.3",
     "@next/plugin-sentry": "^9.5.3",
     "@sentry/webpack-plugin": "^1.13.0",
     "@svgr/webpack": "^5.4.0",
-    "clsx": "^1.1.1",
     "content-security-policy-builder": "^2.1.0",
     "contentful": "^7.14.6",
     "file-loader": "^6.1.0",
