+++ src/components/Image/index.ts
