+++ src/components/SimpleWindow/index.ts
@@ -1,0 +1,1 @@
+export { default } from './SimpleWindow'
