--- src/components/atoms/Logo/__tests__/Logo.tsx
@@ -1,20 +1,0 @@
-import { render, screen } from '@testing-library/react'
-import React from 'react'
-
-import Logo from '../Logo'
-
-import '@testing-library/jest-dom/extend-expect'
-
-describe('Logo', () => {
-  it('basic', () => {
-    render(<Logo />)
-
-    expect(screen.getByTestId('svg-mock')).not.toHaveClass('logo--vertical')
-  })
-
-  it('vertical attribute is given', () => {
-    render(<Logo vertical />)
-
-    expect(screen.getByTestId('svg-mock')).toHaveClass('logo--vertical')
-  })
-})
