--- src/components/pages/Fortune/index.ts
@@ -1,1 +1,0 @@
-export { default } from './Fortune'
