--- src/components/pages/Lottery/index.ts
@@ -1,1 +1,0 @@
-export { default } from './Lottery'
