--- src/components/templates/SingleDoc/index.ts
@@ -1,1 +1,0 @@
-export { default } from './SingleDoc'
