+++ src/pages/_app.tsx
@@ -1,17 +1,50 @@
-import { MDXProvider } from '@mdx-js/react'
+import { cache } from '@emotion/css'
+import { CacheProvider, Global, css } from '@emotion/react'
 import { AppProps, NextWebVitalsMetric } from 'next/app'
-import Head from 'next/head'
 import { DefaultSeo, LogoJsonLd } from 'next-seo'
 import React, { FC } from 'react'
 
-import Link from '@/components/atoms/Link'
-import Progress from '@/components/atoms/Progress'
-import Layout from '@/components/templates/Layout'
+import NProgress from '@/components/NProgress'
 import fullPath from '@/helpers/fullPath'
 
-const MDXComponents = {
-  a: Link
-}
+const globalStyles = css`
+  :root {
+    --ij-default-font-family: -apple-system, BlinkMacSystemFont, Helvetica Neue,
+      Helvetica, Arial, YuGothic, Yu Gothic, sans-serif;
+    --ij-serif-font-family: Garamond, Times New Roman, YuMincho, Yu Mincho,
+      serif;
+  }
+
+  html {
+    box-sizing: border-box;
+    font-family: var(--ij-default-font-family);
+    -webkit-font-smoothing: antialiased;
+    -moz-osx-font-smoothing: grayscale;
+    -webkit-text-size-adjust: none;
+  }
+
+  body {
+    margin: 0;
+    line-height: 2;
+  }
+
+  *,
+  *::before,
+  *::after {
+    box-sizing: inherit;
+  }
+
+  body,
+  html,
+  #__next {
+    height: 100%;
+  }
+
+  p {
+    margin: 0;
+    padding: 0;
+  }
+`
 
 export function reportWebVitals(metric: NextWebVitalsMetric): void {
   const value = metric.name === 'CLS' ? metric.value * 1000 : metric.value
@@ -44,75 +77,13 @@ const MyApp: FC<AppProps> = ({ Component, pageProps }) => (
       url={fullPath('/')}
     />
 
-    <Head>
-      <meta content="#ff5722" name="theme-color" />
+    <CacheProvider value={cache}>
+      <Global styles={globalStyles} />
 
-      <link href="/manifest.webmanifest" rel="manifest" />
-      <link
-        href="/images/favicon-192x192.png"
-        rel="icon"
-        sizes="192x192"
-        type="image/png"
-      />
-      <link
-        href="/images/favicon-512x512.png"
-        rel="icon"
-        sizes="512x512"
-        type="image/png"
-      />
-      <link
-        href="/apple-touch-icon.png"
-        rel="apple-touch-icon"
-        sizes="152x152"
-      />
-    </Head>
+      <Component {...pageProps} />
 
-    <MDXProvider components={MDXComponents}>
-      <Layout>
-        <Component {...pageProps} />
-      </Layout>
-
-      <Progress />
-    </MDXProvider>
-
-    <style global jsx>{`
-      :root {
-        --ij-default-font-family: -apple-system, BlinkMacSystemFont,
-          Helvetica Neue, Helvetica, Arial, YuGothic, Yu Gothic, sans-serif;
-        --ij-serif-font-family: Garamond, Times New Roman, YuMincho, Yu Mincho,
-          serif;
-      }
-
-      html {
-        box-sizing: border-box;
-        font-family: var(--ij-default-font-family);
-        -webkit-font-smoothing: antialiased;
-        -moz-osx-font-smoothing: grayscale;
-        -webkit-text-size-adjust: none;
-      }
-
-      body {
-        margin: 0;
-        line-height: 2;
-      }
-
-      *,
-      *::before,
-      *::after {
-        box-sizing: inherit;
-      }
-
-      body,
-      html,
-      #__next {
-        height: 100%;
-      }
-
-      p {
-        margin: 0;
-        padding: 0;
-      }
-    `}</style>
+      <NProgress />
+    </CacheProvider>
   </>
 )
