+++ src/pages/_document.tsx
@@ -1,16 +1,66 @@
-import Document, { Head, Html, Main, NextScript } from 'next/document'
+import { extractCritical } from '@emotion/server'
+import Document, {
+  DocumentContext,
+  DocumentInitialProps,
+  Head,
+  Html,
+  Main,
+  NextScript
+} from 'next/document'
 import React from 'react'
 
 class MyDocument extends Document {
+  static async getInitialProps(
+    ctx: DocumentContext
+  ): Promise<DocumentInitialProps> {
+    const initialProps = await super.getInitialProps(ctx)
+    const styles = extractCritical(initialProps.html)
+
+    return {
+      ...initialProps,
+      styles: (
+        <>
+          {initialProps.styles}
+          <style
+            data-emotion-css={styles.ids.join(' ')}
+            dangerouslySetInnerHTML={{ __html: styles.css }}
+          />
+        </>
+      )
+    }
+  }
+
   render(): JSX.Element {
     return (
       <Html lang="ja">
         <Head prefix="og: http://ogp.me/ns#">
+          <meta content="#ff5722" name="theme-color" />
+
           <link href="https://www.google-analytics.com" rel="preconnect" />
           <link href="https://www.googletagmanager.com" rel="preconnect" />
+          <link href="/manifest.webmanifest" rel="manifest" />
+          <link
+            href="/images/favicon-192x192.png"
+            rel="icon"
+            sizes="192x192"
+            type="image/png"
+          />
+          <link
+            href="/images/favicon-512x512.png"
+            rel="icon"
+            sizes="512x512"
+            type="image/png"
+          />
+          <link
+            href="/apple-touch-icon.png"
+            rel="apple-touch-icon"
+            sizes="152x152"
+          />
         </Head>
+
         <body>
           <Main />
+
           <NextScript />
         </body>
       </Html>
