+++ src/pages/kuji/[id].tsx
@@ -1,13 +1,44 @@
+import styled from '@emotion/styled'
 import { GetStaticPaths, GetStaticProps, NextPage } from 'next'
 import { BreadcrumbJsonLd, NextSeo } from 'next-seo'
 import React from 'react'
 
-import Fortune from '@/components/pages/Fortune'
+import Image from '@/components/Image'
+import Page from '@/components/Layout'
+import SingleWindow from '@/components/SimpleWindow'
 import getFortune from '@/contentful/getFortune'
 import getFortunes from '@/contentful/getFortunes'
 import fullPath from '@/helpers/fullPath'
+import getTweetLink from '@/helpers/getTweetLink'
 import NotFound from '@/pages/404.mdx'
 import FortuneEntry from '@/types/FortuneEntry'
+
+const ShareLinks = styled.nav`
+  margin-top: 3rem;
+`
+
+const ShareLinksList = styled.ul`
+  align-items: center;
+  display: flex;
+  justify-content: center;
+  list-style-type: none;
+  margin: 0;
+  padding: 0;
+`
+
+const ShareButton = styled.a`
+  background-color: #1da1f2;
+  border: 1px solid #1da1f2;
+  border-radius: 1rem;
+  color: #fff;
+  display: inline-block;
+  font-size: 0.95rem;
+  font-weight: 700;
+  letter-spacing: 0.025em;
+  line-height: 1;
+  padding: 0.5rem 1.5rem;
+  text-decoration: none;
+`
 
 export type Params = {
   id: string
@@ -50,6 +81,8 @@ const KujiPage: NextPage<Props> = ({ fortune }) => {
 
   const name = `第${fortune.fields.number}番『${fortune.fields.blessing}』`
   const title = `因幡はねるくじ ${name}`
+  const imageDetails = fortune.fields.paper.fields.file.details.image
+  const imageURL = `https:${fortune.fields.paper.fields.file.url}`
 
   return (
     <>
@@ -84,7 +117,31 @@ const KujiPage: NextPage<Props> = ({ fortune }) => {
         ]}
       />
 
-      <Fortune fortune={fortune} />
+      <Page>
+        <SingleWindow title={name}>
+          <Image
+            alt={name}
+            height={imageDetails?.height || 0}
+            src={imageURL}
+            width={imageDetails?.width || 0}
+          />
+
+          <ShareLinks>
+            <ShareLinksList>
+              <li>
+                <ShareButton
+                  href={getTweetLink(fortune)}
+                  rel="noopener noreferrer"
+                  role="button"
+                  target="_blank"
+                >
+                  Twitterに共有する
+                </ShareButton>
+              </li>
+            </ShareLinksList>
+          </ShareLinks>
+        </SingleWindow>
+      </Page>
     </>
   )
 }
