+++ src/pages/lottery.tsx
@@ -1,14 +1,35 @@
+import { keyframes } from '@emotion/react'
+import styled from '@emotion/styled'
 import { GetServerSideProps, NextPage } from 'next'
 import Head from 'next/head'
 import { useRouter } from 'next/router'
 import { NextSeo } from 'next-seo'
 import React, { useEffect } from 'react'
 
-import Lottery from '@/components/pages/Lottery'
+import kujiImage from '@/assets/kuji.png'
+import Page from '@/components/Layout'
+import SimpleWindow from '@/components/SimpleWindow'
 import getFortunes from '@/contentful/getFortunes'
 import fullPath from '@/helpers/fullPath'
 
 const DELAY_SECONDS = 2
+
+const shake = keyframes`
+  from {
+    transform: translateY(0) rotate(180deg);
+  }
+
+  to {
+    transform: translateY(20px) rotate(170deg);
+  }
+`
+
+const LotteryBox = styled.img`
+  animation: ${shake} 0.3s infinite alternate linear;
+  display: block;
+  margin: 0 auto;
+  transform: translateY(0) rotate(180deg);
+`
 
 type Props = {
   id: string
@@ -47,6 +68,8 @@ const LotteryPage: NextPage<Props> = ({ id }) => {
       <NextSeo noindex title="おみくじを引いています..." />
 
       <Head>
+        <link as="image" href={kujiImage} rel="preload" />
+
         <noscript>
           <meta
             content={`${DELAY_SECONDS};URL=${fullPath(`/kuji/${id}`)}`}
@@ -55,7 +78,16 @@ const LotteryPage: NextPage<Props> = ({ id }) => {
         </noscript>
       </Head>
 
-      <Lottery />
+      <Page>
+        <SimpleWindow>
+          <LotteryBox
+            alt="くじ引き中..."
+            height="290"
+            src={kujiImage}
+            width="225"
+          />
+        </SimpleWindow>
+      </Page>
     </>
   )
 }
